adds and tests remaining tokens
